create control panel component
